mcp-servers/browser-automation/requirements.txt
# Install from official Tuya MCP SDK repository
# 1. Clone: git clone https://github.com/tuya/tuya-mcp-sdk.git
# 2. cd tuya-mcp-sdk/mcp-python
# 3. pip install -e .

# Additional dependencies
python-dotenv==1.0.0
requests==2.31.0
